get rid of FailureTransportLocator
