+++ FailedMessage/FailedMessageRejecter.php
@@ -3,31 +3,33 @@
 namespace KaroIO\MessengerMonitorBundle\FailedMessage;
 
 use KaroIO\MessengerMonitorBundle\Exception\FailureTransportNotListable;
-use KaroIO\MessengerMonitorBundle\Locator\FailureTransportLocator;
+use KaroIO\MessengerMonitorBundle\Locator\ReceiverLocator;
 use Symfony\Component\Messenger\Transport\Receiver\ListableReceiverInterface;
 
 class FailedMessageRejecter
 {
-    private $failureTransportLocator;
+    private $receiverLocator;
+    private $failureReceiverName;
 
-    public function __construct(FailureTransportLocator $failureTransportLocator)
+    public function __construct(ReceiverLocator $receiverLocator, string $failureReceiverName)
     {
-        $this->failureTransportLocator = $failureTransportLocator;
+        $this->receiverLocator = $receiverLocator;
+        $this->failureReceiverName = $failureReceiverName;
     }
 
     public function rejectFailedMessage($id): void
     {
-        $failureTransport = $this->failureTransportLocator->getFailureTransport();
-        if (!$failureTransport instanceof ListableReceiverInterface) {
+        $failureReceiver = $this->receiverLocator->getReceiver($this->failureReceiverName);
+        if (!$failureReceiver instanceof ListableReceiverInterface) {
             throw new FailureTransportNotListable();
         }
 
-        $envelope = $failureTransport->find($id);
+        $envelope = $failureReceiver->find($id);
 
         if (null === $envelope) {
             throw new \RuntimeException(sprintf('The message with id "%s" was not found.', $id));
         }
 
-        $failureTransport->reject($envelope);
+        $failureReceiver->reject($envelope);
     }
 }
