+++ FailedMessage/FailedMessageRepository.php
@@ -4,6 +4,7 @@ namespace KaroIO\MessengerMonitorBundle\FailedMessage;
 
 use KaroIO\MessengerMonitorBundle\Exception\FailureTransportNotListable;
 use KaroIO\MessengerMonitorBundle\Locator\FailureTransportLocator;
+use KaroIO\MessengerMonitorBundle\Locator\ReceiverLocator;
 use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Stamp\RedeliveryStamp;
 use Symfony\Component\Messenger\Stamp\TransportMessageIdStamp;
@@ -13,22 +14,24 @@ use Symfony\Component\Messenger\Transport\Receiver\ListableReceiverInterface;
 // todo: find a better name?
 class FailedMessageRepository
 {
-    private $failureTransportLocator;
+    private $receiverLocator;
+    private $failureReceiverName;
 
-    public function __construct(FailureTransportLocator $failureTransportLocator)
+    public function __construct(ReceiverLocator $receiverLocator, string $failureReceiverName)
     {
-        $this->failureTransportLocator = $failureTransportLocator;
+        $this->receiverLocator = $receiverLocator;
+        $this->failureReceiverName = $failureReceiverName;
     }
 
     public function listFailedMessages(): array
     {
-        $failureTransport = $this->failureTransportLocator->getFailureTransport();
-        if (!$failureTransport instanceof ListableReceiverInterface) {
+        $failureReceiver = $this->receiverLocator->getReceiver($this->failureReceiverName);
+        if (!$failureReceiver instanceof ListableReceiverInterface) {
             throw new FailureTransportNotListable();
         }
 
         // todo: this number should be dynamic
-        $envelopes = $failureTransport->all(10);
+        $envelopes = $failureReceiver->all(10);
 
         $rows = [];
         foreach ($envelopes as $envelope) {
