+++ Resources/config/services.xml
@@ -33,11 +33,6 @@
             <argument /> <!-- Receiver names -->
         </service>
 
-        <service id="karo-io.messenger_monitor.failure_transport_locator" class="KaroIO\MessengerMonitorBundle\Locator\FailureTransportLocator">
-            <argument type="service" id="messenger.receiver_locator" />
-            <argument /> <!-- Failure transport name -->
-        </service>
-
         <!-- Commands -->
         <service id="karo-io.messenger_monitor.monitor_command" class="KaroIO\MessengerMonitorBundle\Command\MonitorCommand">
             <argument>messenger:monitor</argument>
@@ -47,18 +42,21 @@
 
         <!-- Failed Message -->
         <service id="karo-io.messenger_monitor.failed_message.repository" class="KaroIO\MessengerMonitorBundle\FailedMessage\FailedMessageRepository">
-            <argument type="service" id="karo-io.messenger_monitor.failure_transport_locator" />
+            <argument type="service" id="karo-io.messenger_monitor.receiver_locator" />
+            <argument>%karo-io.messenger_monitor.failure_transport%</argument>
         </service>
 
         <service id="karo-io.messenger_monitor.failed_message.retry" class="KaroIO\MessengerMonitorBundle\FailedMessage\FailedMessageRetryer">
-            <argument type="service" id="karo-io.messenger_monitor.failure_transport_locator" />
+            <argument type="service" id="karo-io.messenger_monitor.receiver_locator" />
+            <argument>%karo-io.messenger_monitor.failure_transport%</argument>
             <argument type="service" id="messenger.routable_message_bus" />
             <argument type="service" id="event_dispatcher" />
             <argument type="service" id="logger" />
         </service>
 
         <service id="karo-io.messenger_monitor.failed_message.reject" class="KaroIO\MessengerMonitorBundle\FailedMessage\FailedMessageRejecter">
-            <argument type="service" id="karo-io.messenger_monitor.failure_transport_locator" />
+            <argument type="service" id="karo-io.messenger_monitor.receiver_locator" />
+            <argument>%karo-io.messenger_monitor.failure_transport%</argument>
         </service>
     </services>
 </container>
